# api/openapi/swagger.yaml
openapi: 3.0.3
info:
  title: Embrace Chaos API
  description: |
    Enterprise-grade chaos engineering platform API for orchestrating chaos experiments 
    across multi-cloud and on-premises infrastructure.
    
    ## Features
    - Multi-cloud provider support (AWS, GCP, Azure, Kubernetes)
    - GitOps workflow integration
    - Real-time safety monitoring
    - Comprehensive audit logging
    - Advanced scheduling and automation
    
    ## Authentication
    This API uses Bearer token authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer your-api-token
    ```
  version: 1.0.0
  contact:
    name: Embrace Chaos API Support
    url: https://github.com/embrace-chaos/embrace-chaos
    email: support@embrace-chaos.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.embrace-chaos.io/v1
    description: Production server
  - url: https://staging-api.embrace-chaos.io/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
paths:
...
  /executions:
    get:
      summary: List executions
      description: Retrieve a paginated list of experiment executions
      operationId: listExecutions
      tags:
        - Executions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/OrderByParam'
        - $ref: '#/components/parameters/OrderDirParam'
        - name: experiment_id
          in: query
          description: Filter by experiment ID
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by execution status
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ExecutionStatus'
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /executions/{executionId}:
    get:
      summary: Get execution
      description: Retrieve a specific execution by ID
      operationId: getExecution
      tags:
        - Executions
      parameters:
        - $ref: '#/components/parameters/ExecutionIdParam'
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /executions/{executionId}/cancel:
    post:
      summary: Cancel execution
      description: Cancel a running experiment execution
      operationId: cancelExecution
      tags:
        - Executions
      parameters:
        - $ref: '#/components/parameters/ExecutionIdParam'
      responses:
        '200':
          description: Execution cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /executions/{executionId}/logs:
    get:
      summary: Get execution logs
      description: Retrieve logs for a specific execution
      operationId: getExecutionLogs
      tags:
        - Executions
      parameters:
        - $ref: '#/components/parameters/ExecutionIdParam'
        - name: follow
          in: query
          description: Follow logs in real-time
          schema:
            type: boolean
            default: false
        - name: tail
          in: query
          description: Number of lines to show from the end
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 100
      responses:
        '200':
          description: Execution logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogResponse'
            text/plain:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Targets endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ExecutionIdParam:
      name: executionId
      in: path
      required: true
      description: Execution ID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OrderByParam:
      name: order_by
      in: query
      description: Field to order by
      schema:
        type: string
        default: created_at
    OrderDirParam:
      name: order_dir
      in: query
      description: Order direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  schemas:
...
    Execution:
      type: object
      required:
        - id
        - experiment_id
        - status
        - started_at
        - trigger_type
        - trigger_by
      properties:
        id:
          type: string
          format: uuid
        experiment_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ExecutionStatus'
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration:
          type: string
          description: Execution duration
        trigger_type:
          type: string
          enum: [manual, scheduled, api, webhook]
        trigger_by:
          type: string
          description: Trigger identifier
        parameters:
          type: object
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
    ExecutionStatus:
      type: string
      enum:
        - pending
        - running
        - succeeded
        - failed
        - cancelled
        - timeout
      description: Execution status

    # Request/Response schemas
...
    ExecutionListResponse:
      type: object
      properties:
        executions:
          type: array
          items:
            $ref: '#/components/schemas/Execution'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # Health schemas
...
    LogResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        total:
          type: integer
    LogEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [debug, info, warn, error]
        message:
          type: string
        source:
          type: string
        metadata:
          type: object

    # Error schemas
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Error details
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Request tracking ID
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - BearerAuth: []
...